+++ src/cpp_src/utils/Election.h
@@ -5,14 +5,14 @@
 class Election {
   public:
     template <typename ProjectsT>
-    Election(long long budget, long long numVoters, ProjectsT &&projects)
+    Election(long long budget, int numVoters, ProjectsT &&projects)
         : budget_(budget), numVoters_(numVoters), projects_(std::forward<ProjectsT>(projects)) {}
     long long budget() const { return budget_; }
-    long long numVoters() const { return numVoters_; }
+    int numVoters() const { return numVoters_; }
     const std::vector<ProjectEmbedding> &projects() const { return projects_; };
 
   private:
     long long budget_;
-    long long numVoters_;
+    int numVoters_;
     std::vector<ProjectEmbedding> projects_;
 };
